+++ graft/coreth/RELEASES.md
@@ -13,6 +13,7 @@
 - Enable expermiental `state-scheme` flag to specify Firewood as a state database.
 - Added prometheus metrics for Firewood if it is enabled and expensive metrics are being used.
 - Disable incompatible APIs for Firewood.
+- Removed deprecated flags `coreth-admin-api-enabled`, `coreth-admin-api-dir`, `tx-regossip-frequency`, `tx-lookup-limit`. Use `admin-api-enabled`, `admin-api-dir`, `regossip-frequency`, `transaction-history` instead.
 
 ## [v0.15.2](https://github.com/ava-labs/coreth/releases/tag/v0.15.2)
 
@@ -22,6 +23,7 @@
 - Major refactor to use [`libevm`](https://github.com/ava-labs/libevm) for EVM execution, database access, types & chain configuration. This improves maintainability and enables keeping up with upstream changes more easily.
 - Add metrics for ACP-176
 - Removed the `"price-options-max-base-fee"` config flag
+- Removed extra type support in "ethclient.BlockByHash", "ethclient.BlockByNumber".
 - Moved extra types returned in `ethclient` package to a new package `plugin/evm/customethclient` which supports the same functionality as `ethclient` but with the new types registered in header and block.
 
 ## [v0.15.0](https://github.com/ava-labs/coreth/releases/tag/v0.15.0)